changed artwork.price to int

+++ artworks/models.py
@@ -5,7 +5,7 @@ from artists.models import Artist
 class Artwork(models.Model):
 	artist = models.ForeignKey(Artist, on_delete=models.DO_NOTHING)
 	title = models.CharField(max_length =75)
-	price = models.DecimalField()
+	price = models.IntegerField()
 	is_sold = models.BooleanField(default=False)
 	listing_date = models.DateTimeField(default=datetime.now, blank=True)
 	description = models.TextField(blank=True)
